requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
djangorestframework-simplejwt==5.3.0 
psycopg2-binary==2.9.10
celery==5.3.4 
redis==5.0.1 
django-celery-beat==2.5.0 
reportlab==4.0.4 
PyPDF2==3.0.1 
python-docx==1.1.0 
pdfplumber==0.10.3 
nltk==3.8.1
requests==2.31.0 
pillow==11.3.0 
whitenoise==6.6.0 
gunicorn==21.2.0 
coverage==7.3.2 
python-dotenv==1.0.0 
django-filter==23.3